Exchange table is created and waiting for data
I keep in mind that it'll be nice to make the table component reusable with some modifications and working with props, but the clock is running. Ill do it if we do well with the schedule, but i have more interesting ideas which i want to carried out first.

I made some cleaning again and some other small improvements.